Refactor dendritic config.
