+++ projects/meta_cl/networks/__init__.py
@@ -20,3 +20,4 @@
 # ----------------------------------------------------------------------
 
 from .anml_networks import ANMLNetwork
+from .dendritic_networks import DendriticNetwork
